Resolve the conflict with the P3323R1 atomic_ref changes by keeping
min/max in __atomic_impl, and align the new code with main:
fetch_max/fetch_min are now constrained to non-const atomic_ref and
return remove_cv_t<_Tp>, the free functions use _CCCL_HOST_DEVICE_API,
and the tests use TEST_HOST_DEVICE_FUNC and UNSUPPORTED: force-tile.

+++ cub/cub/detail/warpspeed/squad/squad.cuh
@@ -47,12 +47,12 @@ struct Squad : SquadDesc
 
   [[nodiscard]] _CCCL_DEVICE_API int warpRank() const
   {
-    return mSpecialRegisters.warpIdx % this->warpCount();
+    return static_cast<int>(mSpecialRegisters.warpIdx % this->warpCount());
   }
 
   [[nodiscard]] _CCCL_DEVICE_API int threadRank() const
   {
-    return mSpecialRegisters.threadIdxX % this->threadCount();
+    return static_cast<int>(mSpecialRegisters.threadIdxX % this->threadCount());
   }
 
   [[nodiscard]] _CCCL_DEVICE_API bool isLeaderThread() const
